backend/app_src/pyproject.toml
[tool.poetry]
name = "stock app"
version = "0.1.0"
description = "Personal stock tracking app"
authors = ["Josh Evans <88986229+Jevans1441@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.7.6"
fastapi = "^0.73.0"
uvicorn = "^0.17.5"
SQLAlchemy = "^1.4.31"
asyncpg = "^0.25.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
black = "^22.1.0"
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
pytest = "^7.0.1"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^3.0.0"
requests = "^2.27.1"
pytest-env = "^0.6.2"
pytest-asyncio = "^0.18.1"
httpx = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
